# app/swagger.yaml
swagger: "2.0"
info:
  title: em-photographe API documentation
  description: Contains all available API endpoints in this codebase
  version: "1.0.0"
  termsOfService: "http://swagger.io/terms/"
host: localhost:3000
basePath: /
schemes:
  - https
paths:
...
  /users/{id}:

    # GET /users/{id} => Read User
    get:
      security:
        - Bearer: []
      tags:
        - User Module
      summary: Read User
      description: API to Read a User
      parameters:
        - name: id
          type: number
          in: path
          description: Id of the User
          required: true
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/UserRes"
        "404":
          description: Not Found
        "500":
          description: Internal Server Error

    # PUT /users/{id} => Update User
    put:
      security:
        - Bearer: []
      tags:
        - User Module
      summary: Update User
      description: API to Update a User
      consumes:
        - multipart/form-data
      parameters:
        - name: id
          type: number
          in: path
          description: Id of the User
          required: true
        - in: body
          name: body
          description: updateUser() Payload
          required: true
          schema:
            $ref: "#/definitions/UserReq"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/BasicRes"
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error

    # DELETE /users/{id} => Delete User
    delete:
      security:
        - Bearer: []
      tags:
        - User Module
      summary: Delete User
      description: API to Delete a User
      parameters:
        - name: id
          type: number
          in: path
          description: Id of the User
          required: true
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: "#/definitions/BasicRes"
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error

  # POST /users/message => Send Message
...
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
definitions:
...
  UserReq:
    properties:
      name:
        type: string
        description: user name
      email:
        type: string
        description: user email
      image:
        type: string
        description: user image
      password:
        type: string
        description: user password

  # ! RESPONSES

  # ARTICLE RESPONSE
...
  BasicRes:
    properties:
      status:
        type: integer
        description: status code
      message:
        type: string
        description: confirm message

  # GALLERY RESPONSE
...
  UserRes:
    type: object
    properties:
      id:
        type: number
        description: user id
      name:
        type: string
        description: user name
      email:
        type: string
        description: user email
      image:
        type: string
        description: user image
      role:
        type: string
        description: user role
      createdAt:
        type: string
        format: date
        description: user creation
      upatedAt:
        type: string
        format: date
        description: user update